+++ ultimate/testData/spring/core/rename/annotationArgBySpELRefInXMLConfig/after/bean/JavaConfig.java
@@ -1,0 +1,9 @@
+package bean;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JavaConfig {
+    @Bean(name = {"nameJ"}) public BeanA createBeanA() { return new BeanA(); }
+}