+++ ultimate/testData/spring/core/rename/annotationArgBySpELRefInXMLConfig/before/bean/KotlinRef.kt
@@ -1,0 +1,8 @@
+package bean
+
+import org.springframework.beans.factory.annotation.Value
+
+class KotlinRef {
+    @Value("#{ nameJ }") private val valueA: BeanA? = null
+    @Value("#{ nameK }") private val valueK: BeanA? = null
+}